* add another form to call filter: 	dst filter(src, filterer, width, height); * add NullKernel for debugging. * add BoxKernel (tested).
git-svn-id: file:///svn/cmlib/trunk@114 25493b3f-faf5-0310-b690-0363d7aa3cd8

+++ include/gil/ext/Scaling.h
@@ -11,7 +11,7 @@ namespace std {
 	}
 
 	template<typename I, typename F>
-	I scale(const I &src, const F &filter, int width, int height)
+	I scale(const I &src, const F &filter, size_t width, size_t height)
 	{
 		I dst(width, height);
 		scale(dst, src, filter);
